Layout update

+++ src/app/exam-register/exam-register-page/exam-register-page.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { DataTransformService } from '../../shared/services/data-transform.service';
 import { SidebarMenuComponent } from '../../shared/sidebar-menu/sidebar-menu.component';
@@ -18,6 +18,8 @@ import { Observable, map, startWith } from 'rxjs';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { ActivatedRoute, Router } from '@angular/router';
 import moment from 'moment';
+import { DialogComponent } from '../../shared/dialog/dialog.component';
+import { ConfirmDialogComponent } from '../../shared/confirm-dialog/confirm-dialog.component';
 
 export class MyErrorStateMatcher implements ErrorStateMatcher {
   isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
@@ -29,7 +31,7 @@ export class MyErrorStateMatcher implements ErrorStateMatcher {
 @Component({
   selector: 'app-exam-register-page',
   standalone: true,
-  imports: [ToolbarComponent, SidebarMenuComponent, MatFormFieldModule, MatInputModule, MatSelectModule, MatFormField, MatDatepickerModule, MatNativeDateModule, MatButtonModule, MatButton, ReactiveFormsModule, CommonModule, NgxMaterialTimepickerModule, HttpClientModule, MatAutocompleteModule],
+  imports: [ToolbarComponent, SidebarMenuComponent, MatFormFieldModule, MatInputModule, MatSelectModule, MatFormField, MatDatepickerModule, MatNativeDateModule, MatButtonModule, MatButton, ReactiveFormsModule, CommonModule, NgxMaterialTimepickerModule, HttpClientModule, MatAutocompleteModule, DialogComponent, ConfirmDialogComponent],
   providers: [DataTransformService, DataService],
   templateUrl: './exam-register-page.component.html',
   styleUrl: './exam-register-page.component.scss'
@@ -46,6 +48,9 @@ export class ExamRegisterPageComponent implements OnInit {
   constructor(private dataTransformService: DataTransformService, private titleService: Title, private fb: FormBuilder, private dataService: DataService, private activatedRoute: ActivatedRoute, private router: Router) { 
     this.isEditing = !!this.activatedRoute.snapshot.paramMap.get('id');
   }
+
+  @ViewChild(DialogComponent) dialog!: DialogComponent;
+  @ViewChild(ConfirmDialogComponent) confirmDialog!: ConfirmDialogComponent;
 
   matcher = new MyErrorStateMatcher()
 
@@ -150,7 +155,7 @@ export class ExamRegisterPageComponent implements OnInit {
           }
           });
           } else {
-            window.alert('Preencha todos os campos obrigatórios corretamente.')
+            this.dialog.openDialog('Preencha todos os campos obrigatórios corretamente.');
     }
   }  
 
@@ -171,31 +176,43 @@ export class ExamRegisterPageComponent implements OnInit {
   
       this.dataService.editData('exams', this.examId, exam).subscribe(() => {
         this.showMessage = true;
+
         this.examRegister.disable();
         this.saveDisabled = true;
+        
   
         setTimeout(() => {
           this.showMessage = false;
         }, 1000);
-  
       });
+
     } else {
-      window.alert('Preencha todos os campos obrigatórios corretamente.');
+      this.dialog.openDialog('Preencha todos os campos obrigatórios corretamente.');
     }
   }
 
   editExam(){
     this.examRegister.enable();
+
+    this.examRegister.get('idPatient')!.disable();
+    this.examRegister.get('name')!.disable();
+
     this.saveDisabled = false;
   }
 
   deleteExam(){
-    this.dataService.deleteData('exams', this.examId).subscribe(() => {
-      window.alert('O registro foi excluído.');
-      this.router.navigate(['/lista-prontuarios']);
-    });
-  }
-  
+    this.confirmDialog.openDialog("Tem certeza que deseja excluir o exame? Essa ação não pode ser desfeita.")
+    const subscription = this.confirmDialog.confirm.subscribe(result => {
+      if (result) {
+        this.dataService.deleteData('exams', this.examId).subscribe(() => {
+          this.router.navigate(['/lista-prontuarios']);
+          subscription.unsubscribe();
+        });
+      } else {
+        subscription.unsubscribe();
+        }
+      });
+    }
 }
