+++ src/app/medical-appoint-reg/medical-appointment-reg-page/medical-appointment-reg-page.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { DataTransformService } from '../../shared/services/data-transform.service';
 import { SidebarMenuComponent } from '../../shared/sidebar-menu/sidebar-menu.component';
@@ -17,6 +17,8 @@ import { Observable, map, startWith } from 'rxjs';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { ActivatedRoute, Router } from '@angular/router';
 import moment from 'moment';
+import { DialogComponent } from '../../shared/dialog/dialog.component';
+import { ConfirmDialogComponent } from '../../shared/confirm-dialog/confirm-dialog.component';
 
 export class MyErrorStateMatcher implements ErrorStateMatcher {
   isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
@@ -28,7 +30,7 @@ export class MyErrorStateMatcher implements ErrorStateMatcher {
 @Component({
   selector: 'app-medical-appointment-reg-page',
   standalone: true,
-  imports: [ToolbarComponent, SidebarMenuComponent, MatFormFieldModule, MatInputModule, MatSelectModule, MatFormField, MatButtonModule, MatButton, ReactiveFormsModule, CommonModule, NgxMaterialTimepickerModule, HttpClientModule, MatAutocompleteModule],
+  imports: [ToolbarComponent, SidebarMenuComponent, MatFormFieldModule, MatInputModule, MatSelectModule, MatFormField, MatButtonModule, MatButton, ReactiveFormsModule, CommonModule, NgxMaterialTimepickerModule, HttpClientModule, MatAutocompleteModule, DialogComponent, ConfirmDialogComponent],
   providers: [DataTransformService, DataService],
   templateUrl: './medical-appointment-reg-page.component.html',
   styleUrl: './medical-appointment-reg-page.component.scss'
@@ -45,6 +47,9 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
   constructor(private dataTransformService: DataTransformService, private titleService: Title, private fb: FormBuilder, private dataService: DataService, private activatedRoute: ActivatedRoute, private router: Router) {
     this.isEditing = !!this.activatedRoute.snapshot.paramMap.get('id')
    }
+
+   @ViewChild(DialogComponent) dialog!: DialogComponent;
+   @ViewChild(ConfirmDialogComponent) confirmDialog!: ConfirmDialogComponent;
 
    matcher = new MyErrorStateMatcher()
   
@@ -146,7 +151,7 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
   }
     });
   } else {
-    window.alert('Preencha todos os campos obrigatórios corretamente.')
+    this.dialog.openDialog('Preencha todos os campos obrigatórios corretamente.');
 
     }
   }  
@@ -176,23 +181,30 @@ export class MedicalAppointmentRegPageComponent implements OnInit {
   
       });
     } else {
-      window.alert('Preencha todos os campos obrigatórios corretamente.');
+      this.dialog.openDialog('Preencha todos os campos obrigatórios corretamente.');
     }
   }
 
   editAppoint(){
     this.appointRegistration.enable();
+
+    this.appointRegistration.get('idPatient')!.disable();
+    this.appointRegistration.get('name')!.disable();
+
     this.saveDisabled = false;
   }
 
   deleteAppoint(){
-    this.dataService.deleteData('appointments', this.appointmentId).subscribe(() => {
-      window.alert('O registro foi excluído.');
-      this.router.navigate(['/lista-prontuarios']);
+    this.confirmDialog.openDialog("Tem certeza que deseja excluir a consulta? Essa ação não pode ser desfeita.")
+    const subscription = this.confirmDialog.confirm.subscribe(result => {
+      if (result) {
+        this.dataService.deleteData('appointments', this.appointmentId).subscribe(() => {
+          this.router.navigate(['/lista-prontuarios']);
+          subscription.unsubscribe();
+        });
+      } else {
+        subscription.unsubscribe();
+      }
     });
   }
-  
 }
-  
-  
-
